+++ src/Video/Info/Util/MetadataTypeSafeReader.php
@@ -7,7 +7,6 @@ namespace Soluble\MediaTools\Video\Info\Util;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 use Soluble\MediaTools\Video\Exception\UnexpectedMetadataException;
-use Webmozart\Assert\Assert;
 
 class MetadataTypeSafeReader
 {
@@ -33,14 +32,15 @@ class MetadataTypeSafeReader
      */
     public function getKeyIntValue(string $key): int
     {
-        try {
-            Assert::integerish(
-                $this->streamMetadata[$key] ?? '<empty>',
-                "The ffprobe/videoInfo metadata '$key' is expected to be an integer. Got: %s"
+        $value = $this->streamMetadata[$key] ?? '<empty>';
+        if (filter_var($value, FILTER_VALIDATE_INT) === false) {
+            $msg = sprintf(
+                "The ffprobe/videoInfo metadata '$key' is expected to be an integer. Got: %s (%s)",
+                gettype($value),
+                (string) $value
             );
-        } catch (\Throwable $e) {
-            $this->logger->notice($e->getMessage());
-            throw new UnexpectedMetadataException($e->getMessage());
+            $this->logger->notice($msg);
+            throw new UnexpectedMetadataException($msg);
         }
 
         return (int) $this->streamMetadata[$key];
@@ -53,14 +53,15 @@ class MetadataTypeSafeReader
      */
     public function getKeyIntOrNullValue(string $key): ?int
     {
-        try {
-            Assert::nullOrIntegerish(
-                $this->streamMetadata[$key] ?? null,
-                "The ffprobe/videoInfo metadata '$key' is expected to be an integer or null. Got: %s"
+        $value = $this->streamMetadata[$key] ?? null;
+        if ($value !== null && filter_var($value, FILTER_VALIDATE_INT) === false) {
+            $msg = sprintf(
+                "The ffprobe/videoInfo metadata '$key' is expected to be an integer or null. Got: %s (%s)",
+                gettype($value),
+                (string) $value
             );
-        } catch (\Throwable $e) {
-            $this->logger->notice($e->getMessage());
-            throw new UnexpectedMetadataException($e->getMessage());
+            $this->logger->notice($msg);
+            throw new UnexpectedMetadataException($msg);
         }
         if (isset($this->streamMetadata[$key])) {
             return (int) $this->streamMetadata[$key];
@@ -76,14 +77,16 @@ class MetadataTypeSafeReader
      */
     public function getKeyFloatValue(string $key): float
     {
-        try {
-            Assert::numeric(
-                $this->streamMetadata[$key] ?? '<empty>',
-                "The ffprobe/videoInfo metadata '$key' is expected to be a float. Got: %s"
+        $value = $this->streamMetadata[$key] ?? '<empty>';
+        if (!is_numeric($value)) {
+            $msg = sprintf(
+                "The ffprobe/videoInfo metadata '$key' is expected to be a float. Got: %s (%s)",
+                gettype($value),
+                (string) $value
             );
-        } catch (\Throwable $e) {
-            $this->logger->notice($e->getMessage());
-            throw new UnexpectedMetadataException($e->getMessage());
+            $this->logger->notice($msg);
+
+            throw new UnexpectedMetadataException($msg);
         }
 
         return (float) $this->streamMetadata[$key];
@@ -96,15 +99,18 @@ class MetadataTypeSafeReader
      */
     public function getKeyFloatOrNullValue(string $key): ?float
     {
-        try {
-            Assert::nullOrNumeric(
-                $this->streamMetadata[$key] ?? null,
-                "The ffprobe/videoInfo metadata '$key' is expected to be a float or null. Got: %s"
+        $value = $this->streamMetadata[$key] ?? null;
+        if ($value !== null && !is_numeric($value)) {
+            $msg = sprintf(
+                "The ffprobe/videoInfo metadata '$key' is expected to be a float or null. Got: %s (%s)",
+                gettype($value),
+                (string) $value
             );
-        } catch (\Throwable $e) {
-            $this->logger->notice($e->getMessage());
-            throw new UnexpectedMetadataException($e->getMessage());
+            $this->logger->notice($msg);
+
+            throw new UnexpectedMetadataException($msg);
         }
+
         if (isset($this->streamMetadata[$key])) {
             return (float) $this->streamMetadata[$key];
         }
@@ -119,15 +125,17 @@ class MetadataTypeSafeReader
      */
     public function getKeyStringOrNullValue(string $key): ?string
     {
-        try {
-            Assert::nullOrString(
-                $this->streamMetadata[$key] ?? null,
-                "The ffprobe/videoInfo metadata '$key' is expected to be a string or null. Got: %s"
+        $value = $this->streamMetadata[$key] ?? null;
+        if ($value !== null && !is_scalar($value)) {
+            $msg = sprintf(
+                "The ffprobe/videoInfo metadata '$key' is expected to be a string or null. Got: %s (%s)",
+                gettype($value),
+                (string) $value
             );
-        } catch (\Throwable $e) {
-            $this->logger->notice($e->getMessage());
-            throw new UnexpectedMetadataException($e->getMessage());
+            $this->logger->notice($msg);
+            throw new UnexpectedMetadataException($msg);
         }
+
         if (isset($this->streamMetadata[$key])) {
             return (string) $this->streamMetadata[$key];
         }
@@ -142,14 +150,15 @@ class MetadataTypeSafeReader
      */
     public function getKeyStringValue(string $key): string
     {
-        try {
-            Assert::string(
-                $this->streamMetadata[$key] ?? null,
-                "The ffprobe/videoInfo metadata '$key' is expected to be a string. Got: %s"
+        $value = $this->streamMetadata[$key] ?? null;
+        if (!is_scalar($value)) {
+            $msg = sprintf(
+                "The ffprobe/videoInfo metadata '$key' is expected to be a string. Got: %s (%s)",
+                gettype($value),
+                (string) $value
             );
-        } catch (\Throwable $e) {
-            $this->logger->notice($e->getMessage());
-            throw new UnexpectedMetadataException($e->getMessage());
+            $this->logger->notice($msg);
+            throw new UnexpectedMetadataException($msg);
         }
 
         return (string) $this->streamMetadata[$key];